Optimize WarehousesByChannelIdLoader loader and move it to warehouse dataloaders
