+++ saleor/graphql/channel/dataloaders.py
@@ -1,16 +1,10 @@
-from collections import defaultdict
-
 from django.db.models import Exists, OuterRef
 
 from ...channel.models import Channel
 from ...order.models import Order
-from ...shipping.models import ShippingZone
-from ...warehouse.models import Warehouse
 from ..checkout.dataloaders import CheckoutByTokenLoader, CheckoutLineByIdLoader
 from ..core.dataloaders import DataLoader
 from ..order.dataloaders import OrderByIdLoader, OrderLineByIdLoader
-from ..shipping.dataloaders import ShippingZoneByIdLoader
-from ..warehouse.dataloaders import WarehouseByIdLoader
 
 
 class ChannelByIdLoader(DataLoader):
@@ -88,60 +82,3 @@ class ChannelWithHasOrdersByIdLoader(DataLoader):
             .in_bulk(keys)
         )
         return [channels.get(channel_id) for channel_id in keys]
-
-
-class ShippingZonesByChannelIdLoader(DataLoader):
-    context_key = "shippingzone_by_channel"
-
-    def batch_load(self, keys):
-        zone_and_channel_is_pairs = (
-            ShippingZone.objects.using(self.database_connection_name)
-            .filter(channels__id__in=keys)
-            .values_list("pk", "channels__id")
-        )
-        channel_shipping_zone_map = defaultdict(list)
-        for zone_id, channel_id in zone_and_channel_is_pairs:
-            channel_shipping_zone_map[channel_id].append(zone_id)
-
-        def map_shipping_zones(shipping_zones):
-            zone_map = {zone.pk: zone for zone in shipping_zones}
-            return [
-                [zone_map[zone_id] for zone_id in channel_shipping_zone_map[channel_id]]
-                for channel_id in keys
-            ]
-
-        return (
-            ShippingZoneByIdLoader(self.context)
-            .load_many({pk for pk, _ in zone_and_channel_is_pairs})
-            .then(map_shipping_zones)
-        )
-
-
-class WarehousesByChannelIdLoader(DataLoader):
-    context_key = "warehouse_by_channel"
-
-    def batch_load(self, keys):
-        warehouse_and_channel_in_pairs = (
-            Warehouse.objects.using(self.database_connection_name)
-            .filter(channels__id__in=keys)
-            .values_list("pk", "channels__id")
-        )
-        channel_warehouse_map = defaultdict(list)
-        for warehouse_id, channel_id in warehouse_and_channel_in_pairs:
-            channel_warehouse_map[channel_id].append(warehouse_id)
-
-        def map_warehouses(warehouses):
-            warehouse_map = {warehouse.pk: warehouse for warehouse in warehouses}
-            return [
-                [
-                    warehouse_map[warehouse_id]
-                    for warehouse_id in channel_warehouse_map[channel_id]
-                ]
-                for channel_id in keys
-            ]
-
-        return (
-            WarehouseByIdLoader(self.context)
-            .load_many({pk for pk, _ in warehouse_and_channel_in_pairs})
-            .then(map_warehouses)
-        )
